feat: support for setting channel HTTP referrer headers for both HTML5 and JS players
This commit introduces support for setting the HTTP referrer headers per channel.
The referer is the http-referrer set for each channel in the m3u8 playlist.
This feature works for both the HTML5 and JS players.

Updates were made to the player-dialog component to handle the referrer headers for the JS player.
Corresponding changes to the html-video-player component for the HTML5 player.